feat : Add message to encourage login to favorites

+++ src/components/favorites/index.tsx
@@ -46,32 +46,37 @@ const Favorites: React.FC<{docsID:number}> = ({docsID}) => {
     // setShowToast(true);
   }
   const handleDeldData = async () => {
-    const resp = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/favorites?user_id=${session?.user?.id}&docs_id=${docsID}`, {
-      method: 'DELETE',
-      headers: {
-          'Content-Type': 'application/json'
+    if(session){
+      const resp = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/favorites?user_id=${session?.user?.id}&docs_id=${docsID}`, {
+        method: 'DELETE',
+        headers: {
+            'Content-Type': 'application/json'
+        }
+      });
+      
+      const { status, data, msg } = await resp.json();
+      if(status === 200 && data){
+        const [docsIdList, docsList]: [number[], DocsWithLink[]] = setFavoritesData(data);
+        setFaoritesIDList(docsIdList);
+        setFavoritesDocsList(docsList);
       }
-    });
-    
-    const { status, data, msg } = await resp.json();
-    if(status === 200 && data){
-      const [docsIdList, docsList]: [number[], DocsWithLink[]] = setFavoritesData(data);
-      setFaoritesIDList(docsIdList);
-      setFavoritesDocsList(docsList);
+      // setToastMessage(msg);
+      // setShowToast(true);
     }
-    // setToastMessage(msg);
-    // setShowToast(true);
+  }
+
+  const beforeLoginMsg = () => {
+    setToastMessage("로그인 후 사용하세요!");
+    setShowToast(true);
   }
 
   return (
-    session
-    ? <>
-        {favoritesActive
-          ? <StarRateIcon className="cursor-pointer mt-px ml-2 text-yellow-400 transition-transform duration-300 hover:rotate-45 hover:text-yellow-300" onClick={handleDeldData}/>
-          : <StarBorderIcon className="cursor-pointer mt-px ml-2 text-yellow-400 transition-transform duration-300 hover:rotate-45 hover:text-yellow-300" onClick={handleAddData}/>}
-        {showToast && <Toast message={toastMessage} onClose={() => setShowToast(false)} />}
-      </>
-    :<></>
+    <>
+      {favoritesActive
+        ? <StarRateIcon className="cursor-pointer mt-px ml-2 text-yellow-400 transition-transform duration-300 hover:rotate-45 hover:text-yellow-300" onClick={session ? handleDeldData : beforeLoginMsg}/>
+        : <StarBorderIcon className="cursor-pointer mt-px ml-2 text-yellow-400 transition-transform duration-300 hover:rotate-45 hover:text-yellow-300" onClick={session ? handleAddData : beforeLoginMsg}/>}
+      {showToast && <Toast message={toastMessage} onClose={() => setShowToast(false)} />}
+    </>
   );
 };
